Add files via upload
Analysis on telecom company that offers its clients two prepaid plans, Surf and Ultimate. The commercial department wants to know which of the plans brings in more revenue in order to adjust the advertising budget. 
I had to carry out a preliminary analysis of the plans based on a relatively small client selection with data on 500 Megaline clients: who the clients are, where they're from, which plan they use, and the number of calls they made and text messages they sent in 2018. My job was to analyze clients' behavior and determine which prepaid plan brings in more revenue.